Test spec example object.
