+++ src/GraphQL/MuckTom.hs
@@ -166,45 +166,3 @@ instance forall ks is fields.
 
 
 data User = User { name :: Text, age :: Int, id :: Int }
-
-
---- TEST STUFF BELOW
-
--- Alternative might be a sum type with deriving Generic and 0-arity constructors?
-type DogCommand = Enum "DogCommand" '["SIT", "DOWN", "HEEL"]
-
-
-type Dog = Object "Dog" '[Pet]
-  '[ Field "name" Text
-   , Field "nickname" Text
-   , Field "barkVolume" Int
-   , Argument "dogCommand" DogCommand :> Field "doesKnowCommand" Bool
-   , Argument "atOtherHomes" (Maybe Bool) :> Field "isHouseTrained" Bool
-   , Field "owner" Human
-   ]
-
-type Sentient = Interface "Sentient" '[Field "name" Text]
-type Pet = Interface "Pet" '[Field "name" Text]
-
-type Alien = Object "Alien" '[Sentient] [Field "name" Text, Field "homePlanet" Text]
-
-type Human = Object "Human" '[Sentient] '[Field "name" Text]
-
-type CatCommand = Enum "CatCommand" '["JUMP"]
-
-type Cat = Object "Cat" '[Pet]
-  '[ Field "name" Text
-   , Field "nickName" (Maybe Text)
-   , Argument "catCommand" CatCommand :> Field "doesKnowCommand" Bool
-   , Field "meowVolume" Int
-   ]
-
-type CatOrDog = Cat :<|> Dog
-type DogOrHuman = Dog :<|> Human
-type HumanOrAlien = Human :<|> Alien
-
-type QueryRoot = Object "QueryRoot" '[Field "dog" Dog]
-
-
-testDefinition :: ObjectTypeDefinition
-testDefinition = getDefinition @Dog
