+++ tests/TypeTests.hs
@@ -1,7 +1,7 @@
-{-# LANGUAGE TypeApplications, DataKinds #-}
+{-# LANGUAGE TypeApplications, DataKinds, TypeOperators #-}
 module TypeTests where
 
-import Protolude
+import Protolude hiding (Enum)
 
 import Test.Tasty (TestTree)
 import Test.Tasty.Hspec (testSpec, describe, it, shouldBe)
@@ -9,10 +9,57 @@ import Test.Tasty.Hspec (testSpec, describe, it, shouldBe)
 import GraphQL.MuckTom
 import GraphQL.Schema
 
-type FD1 = Field "hello" Int
+-- Examples taken from the spec
+
+-- Alternative might be a sum type with deriving Generic and 0-arity constructors?
+type DogCommand = Enum "DogCommand" '["SIT", "DOWN", "HEEL"]
+
+
+type Dog = Object "Dog" '[Pet]
+  '[ Field "name" Text
+   , Field "nickname" Text
+   , Field "barkVolume" Int
+   , Argument "dogCommand" DogCommand :> Field "doesKnowCommand" Bool
+   , Argument "atOtherHomes" (Maybe Bool) :> Field "isHouseTrained" Bool
+   , Field "owner" Human
+   ]
+
+type Sentient = Interface "Sentient" '[Field "name" Text]
+type Pet = Interface "Pet" '[Field "name" Text]
+
+type Alien = Object "Alien" '[Sentient] [Field "name" Text, Field "homePlanet" Text]
+
+type Human = Object "Human" '[Sentient] '[Field "name" Text]
+
+type CatCommand = Enum "CatCommand" '["JUMP"]
+
+type Cat = Object "Cat" '[Pet]
+  '[ Field "name" Text
+   , Field "nickName" (Maybe Text)
+   , Argument "catCommand" CatCommand :> Field "doesKnowCommand" Bool
+   , Field "meowVolume" Int
+   ]
+
+type CatOrDog = Cat :<|> Dog
+type DogOrHuman = Dog :<|> Human
+type HumanOrAlien = Human :<|> Alien
+
+type QueryRoot = Object "QueryRoot" '[Field "dog" Dog]
+
+
+testDefinition :: ObjectTypeDefinition
+testDefinition = getDefinition @Dog
+
 
 typeTests :: IO TestTree
 typeTests = testSpec "Type" $ do
   describe "Field" $
     it "encodes correctly" $ do
     getFieldDefinition @(Field "hello" Int) `shouldBe` (FieldDefinition (Name "hello") [] (TypeNamed (BuiltinType GInt)))
+  describe "Spec" $
+    it "encodes correctly" $ do
+    getDefinition @Human `shouldBe` (
+      ObjectTypeDefinition (Name "Human")
+        [InterfaceTypeDefinition (Name "Sentient") (NonEmptyList [FieldDefinition (Name "name") [] (TypeNamed (BuiltinType GString))])]
+        (NonEmptyList [FieldDefinition (Name "name") [] (TypeNamed (BuiltinType GString))])
+      )
